refactor: extract seed data from main.py + remove dead TenantMiddleware
- Move seed_default_data, _patch_seed_assignments, _sync_yaml_lessons_to_texts
  from main.py (819 lines) to services/seed.py (main.py now 340 lines)
- Delete middleware/tenant.py — request.state.user was never set, so
  TenantMiddleware always returned None for org context (dead code)
- Delete test_tenant_middleware.py (tested dead code)
- Add TODOS.md from full-site eng review (8 items, 2 done)

via [Happy](https://happy.engineering)
