+++ TODOS.md
@@ -1,0 +1,56 @@
+# TODOS — LingoLeap Eng Review 2026-04-05
+
+Generated by `/plan-eng-review` full site review on staging branch.
+
+---
+
+## P1 — High Impact
+
+### 1. Persist Socratic SessionStore to DB
+- **What**: Replace in-memory `SessionStore` dict with DB-backed storage (add `dialogue_state` JSON column to `LearningSession`)
+- **Why**: Cloud Run redeploy/scale loses all Socratic dialogue progress. Students restart from Q1
+- **How to apply**: Add nullable `dialogue_state: JSON` column to `LearningSession`, modify `SessionStore.get()`/`save()` to read/write DB
+- **Depends on**: Alembic migration (need Young's permission)
+- **Effort**: CC ~30 min
+
+### 2. Frontend test coverage — top 5 components + AuthContext
+- **What**: Write tests for LiveTutor, ComprehensionChat, AssessmentReport, VocabDefinitionMatch, AuthContext
+- **Why**: 7% frontend test coverage (16/218 files). Largest components have zero tests
+- **How to apply**: Create `__tests__/` files for each, test happy path + key interactions
+- **Depends on**: Vitest setup (check if vitest.config exists)
+- **Effort**: CC ~1 hour
+
+### 3. Add pagination to teacher analytics + learning errors
+- **What**: Add LIMIT/offset to `teacher_analytics.py` (7 unbounded .all()) and `learning_errors.py` (7 unbounded .all())
+- **Why**: Queries are classroom-scoped (~30-50 students x 57 lessons = ~1500 rows max). Low urgency for current scale, but will matter with growth
+- **How to apply**: Copy pagination pattern from `learning_sessions.py` (already has limit/offset Query params)
+- **Note**: SQL aggregation approach (extract epoch) breaks SQLite tests. Keep Python-side processing for now
+- **Effort**: CC ~20 min
+
+## P2 — Medium Impact
+
+### ~~4. Extract seed data from main.py~~ ✅ DONE
+- Moved to `services/seed.py`. main.py: 819 → 340 lines
+
+### ~~5. Remove dead TenantMiddleware~~ ✅ DONE
+- Deleted `middleware/tenant.py` + `tests/test_tenant_middleware.py` + removed from main.py
+
+### 6. Split LiveTutor.tsx into sub-components
+- **What**: Extract ParagraphView, TutorControlPanel, FeedbackDisplay from LiveTutor.tsx (1,101 lines)
+- **Why**: Over 1000-line critical threshold. Hard to maintain, hard to test
+- **How to apply**: Co-located sub-components in `reading-steps/live-tutor/` directory
+- **Effort**: CC ~20 min
+
+### 7. Narrow bare `except Exception` catches in AI/learning paths
+- **What**: 55 bare `except Exception` catches across backend, 16 in AI/learning paths
+- **Why**: On closer review, these are well-designed fail-safes — they all log the error and fall back gracefully (e.g., `understood=False`, canned fallback question). Circuit breaker handles repeated failures. Low priority refinement
+- **How to apply**: Replace with specific catches where it adds clarity, but don't break the fail-safe pattern
+- **Effort**: CC ~15 min
+
+## P3 — Low Priority (noted, no action needed now)
+
+### 8. Sync SQLAlchemy on async event loop
+- **What**: 239 sync endpoints block the event loop thread
+- **Why**: FastAPI thread pool (40) handles this, but could exhaust under heavy load
+- **When to act**: When you see slow responses under concurrent load (50+ students)
+- **No action needed now**
